add option -H to specify a colon-separated list of headers to include
ags editor recursively scans a game dir and includes all headers.
how it knows the right order to include them is unknown.
a symbol from one header can depend upon one in another, and having
the order wrong breaks compilation.
in the realm of UNIX and C, we dislike doing implicit wacky things,
so instead of adding 100+ lines of code to do globbing, recursing,
and figuring the order out by ignoring compile errors, we simply
demand from the user that he specifies them himself in the right
order.

example:

    ascc -H ../game/Vehicles.ash:/tmp/test.ash:Timer.ash Timer.asc

just like the filename of the .asc file to be compiled, a full
path must be specified for each item (relative or absolute).
with this change, the open source game "Last n' Furious" can be
successfully compiled.