turn off boundchecks in non-debug mode
by default, AGS compiler used to emit a boundscheck "assertlte"
instruction when a non-dynamic array is accessed, which checks
that the index into the array is smaller than its size.

since it is quite costly (and the engine/VM should check script
memory accesses anyway), we emit it only in debug mode.